[changed] Renames TabbedArea to Tabs and TabPane to Tab
* Deprecates TabbedArea and TabPane
* Adds Tabs (formerly TabbedArea) and Tab (formerly TabPane)
* `tab` attribute is renamed to `title`
* Removes tests for TabbedArea and TabPane (because the deprecation
warning fails them all
* Update docs to use Tabs and Tab examples
