+++ docs/examples/.eslintrc
@@ -47,9 +47,11 @@
     "ProgressBar",
     "Row",
     "SplitButton",
+    "Tab",
     "TabbedArea",
     "Table",
     "TabPane",
+    "Tabs",
     "Tooltip",
     "Well",
     "Thumbnail",
