+++ docs/examples/TabsControlled.js
@@ -1,0 +1,24 @@
+const ControlledTabs = React.createClass({
+  getInitialState() {
+    return {
+      key: 1
+    };
+  },
+
+  handleSelect(key) {
+    alert('selected ' + key);
+    this.setState({key});
+  },
+
+  render() {
+    return (
+      <Tabs activeKey={this.state.key} onSelect={this.handleSelect}>
+        <Tab eventKey={1} title='Tab 1'>Tab 1 content</Tab>
+        <Tab eventKey={2} title='Tab 2'>Tab 2 content</Tab>
+        <Tab eventKey={3} title='Tab 3' disabled>Tab 3 content</Tab>
+      </Tabs>
+    );
+  }
+});
+
+React.render(<ControlledTabs />, mountNode);
