+++ docs/src/ReactPlayground.js
@@ -46,9 +46,11 @@ const Portal = require('../../src/Portal');
 const ProgressBar = require('../../src/ProgressBar');
 const Row = require('../../src/Row');
 const SplitButton = require('../../src/SplitButton');
+const Tab = require('../../src/Tab');
 const TabbedArea = require('../../src/TabbedArea');
 const Table = require('../../src/Table');
 const TabPane = require('../../src/TabPane');
+const Tabs = require('../../src/Tabs');
 const Thumbnail = require('../../src/Thumbnail');
 const Tooltip = require('../../src/Tooltip');
 const Well = require('../../src/Well');
