+++ src/index.js
@@ -53,9 +53,11 @@ export SafeAnchor from './SafeAnchor';
 export SplitButton from './SplitButton';
 export styleMaps from './styleMaps';
 export SubNav from './SubNav';
+export Tab from './Tab';
 export TabbedArea from './TabbedArea';
 export Table from './Table';
 export TabPane from './TabPane';
+export Tabs from './Tabs';
 export Thumbnail from './Thumbnail';
 export Tooltip from './Tooltip';
 export Well from './Well';
